Move autodispose lifecycle artifact to autodispose2.* package
