+++ android/autodispose-android-archcomponents/src/androidTest/java/autodispose2/android/lifecycle/AndroidLifecycleScopeProviderTest.java
@@ -15,20 +15,20 @@
  */
 package autodispose2.android.lifecycle;
 
-import static com.google.common.truth.Truth.assertThat;
-import static autodispose2.AutoDispose.autoDisposable;
-
 import android.util.Log;
 import androidx.lifecycle.Lifecycle;
 import androidx.test.annotation.UiThreadTest;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import autodispose2.android.lifecycle.test.TestLifecycleOwner;
-import com.uber.autodispose.lifecycle.LifecycleEndedException;
+import autodispose2.lifecycle.LifecycleEndedException;
 import com.uber.autodispose.test.RecordingObserver;
 import io.reactivex.rxjava3.disposables.Disposable;
 import io.reactivex.rxjava3.subjects.PublishSubject;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+
+import static autodispose2.AutoDispose.autoDisposable;
+import static com.google.common.truth.Truth.assertThat;
 
 @RunWith(AndroidJUnit4.class)
 public final class AndroidLifecycleScopeProviderTest {
