+++ autodispose-lifecycle/src/main/java/autodispose2/lifecycle/LifecycleScopeProvider.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.uber.autodispose.lifecycle;
+package autodispose2.lifecycle;
 
 import autodispose2.ScopeProvider;
 import autodispose2.internal.DoNotMock;
