+++ autodispose-lifecycle/src/main/java/autodispose2/lifecycle/LifecycleScopes.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.uber.autodispose.lifecycle;
+package autodispose2.lifecycle;
 
 import autodispose2.AutoDisposePlugins;
 import autodispose2.OutsideScopeException;
