+++ autodispose-lifecycle/src/test/java/autodispose2/lifecycle/LifecycleScopesTest.java
@@ -13,10 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.uber.autodispose.lifecycle;
+package autodispose2.lifecycle;
 
-import static com.uber.autodispose.lifecycle.LifecycleScopes.resolveScopeFromLifecycle;
-import static com.uber.autodispose.lifecycle.TestLifecycleScopeProvider.TestLifecycle.STOPPED;
+import static autodispose2.lifecycle.LifecycleScopes.resolveScopeFromLifecycle;
+import static autodispose2.lifecycle.TestLifecycleScopeProvider.TestLifecycle.STOPPED;
 
 import autodispose2.AutoDisposePlugins;
 import io.reactivex.rxjava3.core.CompletableSource;
