+++ sample/src/main/java/com/uber/autodispose/recipes/AutoDisposeActivity.java
@@ -18,9 +18,9 @@ package com.uber.autodispose.recipes;
 import android.app.Activity;
 import android.os.Bundle;
 import androidx.annotation.Nullable;
-import com.uber.autodispose.lifecycle.CorrespondingEventsFunction;
-import com.uber.autodispose.lifecycle.LifecycleEndedException;
-import com.uber.autodispose.lifecycle.LifecycleScopeProvider;
+import autodispose2.lifecycle.CorrespondingEventsFunction;
+import autodispose2.lifecycle.LifecycleEndedException;
+import autodispose2.lifecycle.LifecycleScopeProvider;
 import io.reactivex.rxjava3.core.Observable;
 import io.reactivex.rxjava3.subjects.BehaviorSubject;
