+++ sample/src/main/java/com/uber/autodispose/recipes/AutoDisposeFragment.java
@@ -20,9 +20,9 @@ import android.content.Context;
 import android.os.Bundle;
 import android.view.View;
 import androidx.annotation.Nullable;
-import com.uber.autodispose.lifecycle.CorrespondingEventsFunction;
-import com.uber.autodispose.lifecycle.LifecycleEndedException;
-import com.uber.autodispose.lifecycle.LifecycleScopeProvider;
+import autodispose2.lifecycle.CorrespondingEventsFunction;
+import autodispose2.lifecycle.LifecycleEndedException;
+import autodispose2.lifecycle.LifecycleScopeProvider;
 import io.reactivex.rxjava3.core.Observable;
 import io.reactivex.rxjava3.subjects.BehaviorSubject;
