+++ sample/src/main/kotlin/com/uber/autodispose/recipes/AutoDisposeViewKotlin.kt
@@ -21,9 +21,9 @@ import android.util.AttributeSet
 import android.view.View
 import androidx.annotation.RequiresApi
 import autodispose2.android.ViewScopeProvider
-import com.uber.autodispose.lifecycle.CorrespondingEventsFunction
-import com.uber.autodispose.lifecycle.LifecycleEndedException
-import com.uber.autodispose.lifecycle.LifecycleScopeProvider
+import autodispose2.lifecycle.CorrespondingEventsFunction
+import autodispose2.lifecycle.LifecycleEndedException
+import autodispose2.lifecycle.LifecycleScopeProvider
 import com.uber.autodispose.recipes.AutoDisposeViewKotlin.ViewEvent
 import com.uber.autodispose.recipes.AutoDisposeViewKotlin.ViewEvent.ATTACH
 import com.uber.autodispose.recipes.AutoDisposeViewKotlin.ViewEvent.DETACH
