+++ static-analysis/autodispose-error-prone/src/test/resources/com/uber/autodispose/errorprone/UseAutoDisposeNegativeCasesExcluded.java
@@ -18,10 +18,10 @@ package com.uber.autodispose.errorprone;
 import static autodispose2.AutoDispose.autoDisposable;
 
 import autodispose2.AutoDispose;
-import com.uber.autodispose.lifecycle.CorrespondingEventsFunction;
-import com.uber.autodispose.lifecycle.LifecycleEndedException;
-import com.uber.autodispose.lifecycle.LifecycleScopeProvider;
-import com.uber.autodispose.lifecycle.TestLifecycleScopeProvider.TestLifecycle;
+import autodispose2.lifecycle.CorrespondingEventsFunction;
+import autodispose2.lifecycle.LifecycleEndedException;
+import autodispose2.lifecycle.LifecycleScopeProvider;
+import autodispose2.lifecycle.TestLifecycleScopeProvider.TestLifecycle;
 import io.reactivex.rxjava3.annotations.CheckReturnValue;
 import io.reactivex.rxjava3.annotations.Nullable;
 import io.reactivex.rxjava3.core.Completable;
